Move MethodAndUrlUtils to GraphUtils unit
